PERMISSIONS
Fixed some error logs with permissions
Added checks if the bot is muted or not to a few extra functions

+++ tg_bot/modules/global_bans.py
@@ -700,7 +700,8 @@ def welcome_gban(update, context, user_id):
     if sql.is_user_gbanned(user_id):
         chat.kick_member(user_id)
         try:
-            msg.delete()
+            if msg:
+                msg.delete()
         except:
             LOGGER.warning("Could not find the message to delete.", exc_info=1)
